requirements.txt
hyperliquid-python-sdk>=0.21.0
eth-account
python-dotenv
websockets
